Sync Z studies with CERN

+++ templateMaker/interface/lumiWeight.hpp
@@ -9,10 +9,10 @@ class lumiWeight : public Module
 private:
     float _targetLumi;
     float _xsec;
-    float _genEventSumwClipped;
+    double _genEventSumwClipped;
 
 public:
-  lumiWeight(float xsec, float sumwclipped, float targetLumi = 35.9)
+  lumiWeight(float xsec, double sumwclipped, float targetLumi = 35.9)
     {
         _targetLumi = targetLumi;
         _xsec = xsec/ 0.001;
